+++ pyldap/ldapmodlist.h
@@ -16,14 +16,14 @@
 typedef struct {
 	PyObject_HEAD
 	LDAPMod **mod_list;
-	unsigned short int last;
-	unsigned short int size;
+	Py_ssize_t last;
+	Py_ssize_t size;
 	PyObject *entry;
 } LDAPModList;
 
 extern PyTypeObject LDAPModListType;
 
-LDAPModList *LDAPModList_New(PyObject *entry, unsigned short int size);
+LDAPModList *LDAPModList_New(PyObject *entry, Py_ssize_t size);
 int LDAPModList_Add(LDAPModList *self, int mod_op, PyObject *key, PyObject *value);
 PyObject *LDAPModList_Pop(LDAPModList *self);
 int LDAPModList_Empty(LDAPModList *self);
